docker/requirements.txt
yt-dlp
pydub
noisereduce
SpeechRecognition
mysql-connector-python
ffmpeg-python
numpy
pyttsx3
pyaudio
opencv-python
python-dotenv

